fix: don't reset to regular font when bold or italic is None (#125)
We already set font when first creating the style_builder, no need to reset it. This might also be wiping a previous setting. For example, if it has both italics and bold modifier, italics font is applied, but bold matches none, and resets font back to regular, we want to avoid such a behavior I assume.